requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
bs4==0.0.1
feedparser==5.2.1
html5lib==0.999999999
packaging==16.8
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
six==1.10.0
twitter==1.17.1
webencodings==0.5
